:recycle: Replace Scripts/Stylesheets/Styles helpers
My original vision for loadScript, loadStylesheet, and addStyles no longer aligns with their current use-cases and they need to be refactored.

Some problems with them:
* they are persistent by default, depending on the script or styles to scope themselves to the appropriate pages.
* they are not named consistently, with 'load' and 'add' not wholly describing their purpose.
* they inconsistently use an options property, in which addStyles uses options for persistence and loadScript and loadStylesheet use options to set certain attributes on the element.

New revisions:
* they are all named with 'insert' for consistency
* insertScript and insertStylesheet replace their 'source' and 'options' arguments with a single 'attributes' argument
* they each have a 'shouldPersist' argument. By default any added scripts or styles are destroyed upon navigation
* they tap into the recently-created 'grove-navigation' event

:warning: WARNING! Although styles and stylesheets removed from the DOM also destroy the added styles, removing a script does not stop the script from continuing to run. Any eventListeners or intervals created continue to execute.

:boom: This change is breaking.
A stylesheet loaded with Glade.loadStylesheet(PATH) must now be called as Glade.insertStylesheet({href: PATH}).
A script loaded with Glade.loadScript(PATH) must now be called as Glade.insertScript({src: PATH}).
A script loaded with Glade.loadScript(PATH, {defer: true}) must now be called as Glade.insertScript({src: PATH, defer: true}).
Although those changes are a bit of a nuisance, it's a good tradeoff than having a path argument and a configuration argument, and it also comes with the bonus that all attributes are supported by this script type. It also avoids a pattern of Glade.loadScript(PATH, {}, true) which would occur when a script should persist but does not need custom attributes.

That all script/style loading helpers are not persistent by default is also a breaking change.