:wastebasket: Deprecate STLPR Alias plugin
Originally Glade was the STLPR Object. Once the project was decoupled with the STLPR site, it was renamed to Glade. The STLPR Alias plugin was intended to make this migration a non-breaking change.

I've tracked Grove content I could find which uses Glade and renamed STLPR references to Glade. If there are other pages which use it, I want them to fail loudly.