:construction: Prepare onNavigate helper to fix Glade failing on navigation
Sometimes a workaround would be applied twice or wouldn't appear at all. Upon investigation, I discovered this was related to up to 100 ms of delay that can occur between when a page is navigated to and when the Page Navigation event fires.

Because it can take up to 100ms for the event to fire, a function hooking to the Navigation event within this time window may be executed immediately afterwards when the event fires.

Scripts where a function runs and is then assigned on navigation may execute twice.

Scripts that execute a function and then run cleanup on navigation may have the cleanup executed immediately after the initial function.

The issue came to light August 12, 2021 when all scripts on the staff page broke. This was exposed because a Brightspot update on August 11 improved the page loading speeds considerably, so that it was nearly always finished within that 100ms window.

While setTimeouts are a bit of a code smell, it should be acceptable in this circumstance. We are not using setTimeout to bypass our event stack and we are not repeating our timeouts in a way that could cause them to execute endlessly.

Now that IE is deprecated, though, future solutions should consider some type of listener or observer function for improved performance and decreased risk.