:recycle: Improvements to UMSL Attribution
:speech_balloon: Add UMSL Logo back to hamburger menu credit.
:sparkles: Add UMSL Attribution to the footer
:recycle: Add links to UMSL Logos (2/9/2021)
:sparkles: Remove service-line flag so UMSL Atribution is live (2/10/2021)
:lipstick: Adjust placement of UMSL logo (2/17/2021)
:bee: Fix Persistance bugs for UMSL attribution (2/22/2021)

The behavior of the header and footer atriubtion was inconsistent. Within the header, the text content persisted but the CSS did not. Within the footer, neither the text nor styling persisted.

This is because Grove's Single-Page Application behavior persists the header on navigation, but no other part of the page. The solution is to force all CSS to persist and to repeat addUMSLFooterAttribution on navigation.

:speech_balloon: Replace 'member-supported' with 'listener-supported' language in the Service Line (4/15/2021)
:recycle: Apply new onNavigate helper to correct Glade navigation bugs (11/18/2021)
:recycle: Check the existence of elements such as the footer before using them. (8/3/2022)

A major Grove code push broke a couple of these plugins. Ensure that elements that can cause errors are checked and the error reported.