:recycle: Turn rebranding demo into functional animated branded logos
:sparkle: Go Live with Logo animation (4/15/2021)

I pulled rebrandingFlag.js apart piece by piece. Branding stylesheet is moved into Grove's global stylsheet. Color palette changes are applied in Grove. Grove is updated with the header and footer.

This script is not limited in scope to just swapping out the PNG images in the header/footer and replacing them with animated SVG.

:bee: Correct class names that have changed in Grove (8/3/2022)