:recycle: Add an onNavigation queue for troubleshooting
While testing a popup module, it occurred to me that a navigate event within a function that is reloaded could cause a function to be bound to onNavigate more than once. If a function is bound to the navigate event every time the user navigates to a new page, we could see dozens of the same event being fired and lead to unexpected behavior.

Thankfully, the code I was testing did not create duplicate events, but it highlights the need to see what functions are bound to onNavigate.

My solution is to create a Glade.onNavigate[] property that a *copy* of all functions bound to the navigate event can be viewed on. This means that a user reaching for Glade.onNavigate could execute these functions manually, but removing them from this Array or deleting them would not prevent them from running on the next page