:recycle: Return a destructor function from onNavigate
If a function is bound to the navigate event, there isn't currently an approved method for unbinding it. One could piece together an approach by exploring the Glade.onNavigate queue, but an ideal solution is to return a destructor function.

This is somewhat similar to how insertCSS, insertScript, insertStylesheet, and setIsolatedInterval all return a reference to their respective elements or interval so that the user can manipulate it later.