:recycle: Remove dependence on window.brightspotDataLayer
window.brightspotDataLayer contained page metadata, but had finnicky timing with Glade scripts that could cause scripts to fail on the first load.

A much more reliable approach is to read the brightspot-dataLayer meta tag that exists in pages. This element is guaranteed to exist before any script in Glade is loaded and JSON parsing the object allows us to access all of its properties, even those that are not normally useful.