:recycle: If UMSL Attribution cannot be placed in the header or footer, do not fail immediately
With the previous behaivor, if the Navigation menu container was missing, the script would fail and would not attempt to load the footer attribution.

If the Footer navigation container was missing, the script would fail and would not attempt to load the navigation menu attribution when the user navigates to another page.

To fix this, instead of throwing Errors, log the problem and skip to the next check. This way if there is any context in which the attribution can be loaded, now or in the future, it will be.